+++ core-ng/src/test/java/core/framework/internal/http/RetryInterceptorTest.java
@@ -1,6 +1,8 @@
 package core.framework.internal.http;
 
 import core.framework.api.http.HTTPStatus;
+import core.framework.internal.log.ActionLog;
+import core.framework.internal.log.LogManager;
 import okhttp3.Request;
 import okhttp3.internal.http2.ConnectionShutdownException;
 import okhttp3.internal.http2.ErrorCode;
@@ -419,5 +421,17 @@ class RetryInterceptorTest {
     void uri() {
         var request = new Request.Builder().url("http://localhost/path?query=value").build();
         assertThat(interceptor.uri(request)).isEqualTo("http://localhost/path");
+    }
+
+    @Test
+    void withinMaxProcessTime() {
+        var logManager = new LogManager();
+        ActionLog actionLog = logManager.begin("begin", null);
+
+        actionLog.maxProcessTime(Duration.ofSeconds(1).toNanos());
+        assertThat(interceptor.withinMaxProcessTime(1)).isTrue();
+        assertThat(interceptor.withinMaxProcessTime(2)).isFalse();
+
+        logManager.end("end");
     }
 }