+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 * httpClient: verify hostname with trusting specific cert !!! behavior changed, check carefully
   > previously, HTTPClient.builder().trust(CERT) will use empty hostname verifier
   > now, only trustAll() will bypass hostname verifying, so right now if you trust specific cert, make sure you have CN/ALT_NAME matches the dns name (wildcard domain name also works)
+* httpClient: retry aware of maxProcessTime
+  > for short circuit, e.g. heavy load request causes remote service busy, and client timeout triggers more retry requests, to amplify the load
 
 ### 7.10.5 (04/01/2022 - 04/14/2022)